deploy/librechat-gcp/speech-bridge/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-speech==2.28.0
google-cloud-texttospeech==2.18.0
pydantic==2.9.2
python-multipart==0.0.12
